+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased](https://github.com/CrowdHailer/Ace) - xxxx-xx-xx
+
+## Added
+- System sends welcome message
+
+### Changed
+
 ## [0.3.0](https://github.com/CrowdHailer/Ace) - 2016-10-04
 
 ## Added